fix: separate tab content into screens
