+++ libs/service-portal/health/src/screens/Medicine/Medicine.css.ts
