+++ libs/service-portal/health/src/screens/Medicine/MedicineLicense.tsx
@@ -5,14 +5,14 @@ import {
   Stack,
 } from '@island.is/island-ui/core'
 import { useLocale } from '@island.is/localization'
-import { messages } from '../../../lib/messages'
-import { useGetDrugCertificatesQuery } from '../Medicine.generated'
-import { SECTION_GAP } from '../constants'
+import { messages } from '../../lib/messages'
+import { useGetDrugCertificatesQuery } from './Medicine.generated'
+import { SECTION_GAP } from './constants'
 import { ActionCard, IntroHeader, m } from '@island.is/service-portal/core'
-import { HealthPaths } from '../../../lib/paths'
-import { MedicineWrapper } from '../components/MedicineWrapper/MedicineWrapper'
+import { HealthPaths } from '../../lib/paths'
+import { MedicineWrapper } from './components/MedicineWrapper/MedicineWrapper'
 
-export const MedicineLicence = () => {
+export const MedicineLicense = () => {
   const { formatMessage } = useLocale()
 
   const { data, error, loading } = useGetDrugCertificatesQuery()
@@ -78,3 +78,5 @@ export const MedicineLicence = () => {
     </MedicineWrapper>
   )
 }
+
+export default MedicineLicense
